requirements.txt
Flask>=3.0,<4.0
Flask-Cors>=4.0,<5.0
neo4j>=5.25,<7.0
pdfplumber
python-docx
gunicorn
python-dotenv
